*Changed PetSpellState enum to ushort *Fixed writing of PetSpellState to SMSG_PET_SPELLS

+++ Core/WCell.Constants/Pets/Misc.cs
@@ -52,7 +52,7 @@ namespace WCell.Constants.Pets
 	}
 
 	[Flags]
-	public enum PetSpellState : byte
+	public enum PetSpellState : ushort
 	{
 		None,
 		Passive = 1,
